test: support runnning raw
can be run via something like:
FLOX_BIN=flox NIX_BIN=nix TESTS_DIR=$PWD/cli/tests SHELL=bash PKGDB_BIN=/nix/store/zggaya81gm8m7y8bkjjrx7fxjynmaxaz-flox-pkgdb-0.1.0/bin/pkgdb __FT_RAN_HOME_SETUP=real bats cli/tests

+++ cli/tests/setup_suite.bash
@@ -305,12 +305,12 @@ xdg_tmp_setup() {
   mkdir -p "${XDG_CONFIG_HOME:?}"
   chmod u+w "$XDG_CONFIG_HOME"
 
-  if [[ -e "${REAL_XDG_CONFIG_HOME:?}/nix" ]]; then
+  if [[ -e "${REAL_XDG_CONFIG_HOME:?}/nix" && "$REAL_XDG_CONFIG_HOME" != "$XDG_CONFIG_HOME" ]]; then
     rm -rf "$XDG_CONFIG_HOME/nix"
     cp -Tr -- "$REAL_XDG_CONFIG_HOME/nix" "$XDG_CONFIG_HOME/nix"
     chmod -R u+w "$XDG_CONFIG_HOME/nix"
   fi
-  if [[ -e "$REAL_XDG_CONFIG_HOME/flox" ]]; then
+  if [[ -e "${REAL_XDG_CONFIG_HOME}/flox" && "$REAL_XDG_CONFIG_HOME" != "$XDG_CONFIG_HOME" ]]; then
     rm -rf "$XDG_CONFIG_HOME/flox"
     cp -Tr -- "$REAL_XDG_CONFIG_HOME/flox" "$XDG_CONFIG_HOME/flox"
     chmod -R u+w "$XDG_CONFIG_HOME/flox"
@@ -391,6 +391,11 @@ flox_vars_setup() {
 # Homedirs can be created "globally" for the entire test suite ( default ), or
 # for individual files or single tests by passing an optional argument.
 home_setup() {
+  if [[ "${__FT_RAN_HOME_SETUP:-}" = "real" ]]; then
+    export FLOX_TEST_HOME="$REAL_HOME"
+    export HOME="$REAL_HOME"
+    return 0;
+  fi
   case "${1:-suite}" in
     suite) export FLOX_TEST_HOME="${BATS_SUITE_TMPDIR?}/home" ;;
     file) export FLOX_TEST_HOME="${BATS_FILE_TMPDIR?}/home" ;;
@@ -400,7 +405,6 @@ home_setup() {
       return 1
       ;;
   esac
-  #if [[ "${__FT_RAN_HOME_SETUP:-}" = "$FLOX_TEST_HOME" ]]; then return 0; fi
   # Force recreation on `home' on every invocation.
   unset __FT_RAN_HOME_SETUP
   xdg_tmp_setup
@@ -439,6 +443,7 @@ common_suite_setup() {
   {
     print_var FLOX_TEST_HOME
     print_var HOME
+    print_var PATH
     print_var XDG_CACHE_HOME
     print_var XDG_CONFIG_HOME
     print_var XDG_DATA_HOME
